Refactor code: Move various DLC analysis functions to backend
Now every deeplabcut function is in dlc_helper_functions.py
